requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.0
openai>=1.10.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
youtube-transcript-api>=0.6.1
tiktoken>=0.5.2
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
tenacity>=8.2.3
